bunny_scripts/requirements.txt
# Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Altyazı işleme
pysubs2==1.6.1

# AI çeviri
openai==1.3.5
google-generativeai==0.3.1

# HTTP istekleri
requests==2.31.0
httpx==0.25.1

# Cache
redis==5.0.1

# Env
python-dotenv==1.0.0

# TürkAnime scraping (mevcut)
selenium==4.15.2
